# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Technical Test ",
        "description": "documentation for API Technical Test",
        "contact": {
            "email": "maszahid1@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/technical-test/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Endpoint ini untuk authentication login",
                "operationId": "9bd34eaa85fa47cc53c314cf0b304b5a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "example": "maszahid1@gmail.com"
                                    },
                                    "password": {
                                        "description": "Minimal 7 karakte, harus terdapat huruf kecil dan besar, angka dan simbol",
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "ybrakus@example.net",
                                    "password": "Password2!"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Akan memberikan token authentication, tipe token dan batas expired"
                    },
                    "422": {
                        "description": "Some variable is required"
                    },
                    "401": {
                        "description": "Unathorized"
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Endpoint ini untuk mendapatkan data user yang sedang login",
                "operationId": "7479d1e3fef943aefe451a83dc79b6b4",
                "responses": {
                    "200": {
                        "description": "Akan memberikan data user yang sedang login"
                    },
                    "403": {
                        "description": "Belum melakukan login"
                    },
                    "401": {
                        "description": "Unathorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Endpoint ini untuk logout user yang sedang login",
                "operationId": "caa5de44404b121464ac2da7c9b07744",
                "responses": {
                    "200": {
                        "description": "Akan memberikan pesan berhasil logout"
                    },
                    "403": {
                        "description": "Belum melakukan login"
                    },
                    "401": {
                        "description": "Unathorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/refresh": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Endpoint ini untuk mendapatkan token baru user yang sedang login",
                "operationId": "fc1300d3fa82c41745015d1a0ec67de5",
                "responses": {
                    "200": {
                        "description": "Akan memberikan token baru"
                    },
                    "403": {
                        "description": "Belum melakukan login"
                    },
                    "401": {
                        "description": "Unathorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Endpoint ini untuk registrasi akun baru",
                "operationId": "17736d4f36a850c7df6b6781f5afd6b1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "value harus berupa email address",
                                        "type": "email",
                                        "example": "maszahid1@gmail.com"
                                    },
                                    "password": {
                                        "description": "Minimal 7 karakte, harus terdapat huruf kecil dan besar, angka dan simbol",
                                        "type": "string",
                                        "example": "Abcdefg1!"
                                    },
                                    "password_confirmation": {
                                        "description": "password konfirmasi harus sama",
                                        "type": "string",
                                        "example": "Abcdefg1!"
                                    },
                                    "firstname": {
                                        "description": "nama depan",
                                        "type": "string",
                                        "example": "Alexander"
                                    },
                                    "lastname": {
                                        "description": "nama belakang",
                                        "type": "string",
                                        "example": "Graham Bell"
                                    },
                                    "bank_name": {
                                        "description": "Nama Bank yang akan didaftarkan",
                                        "type": "string",
                                        "example": "BCA/BNI/BRI"
                                    },
                                    "rekening": {
                                        "description": "Nomor Rekening",
                                        "type": "integer",
                                        "example": "57998977"
                                    },
                                    "saldo": {
                                        "description": "Setoran Pertama, minimal Rp50.000",
                                        "type": "integer",
                                        "example": "50000"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "testing@example.net",
                                    "password": "Password2!",
                                    "password_confirmation": "Password2!",
                                    "firstname": "Alexander",
                                    "lastname": "Graham Bell",
                                    "bank_name": "BCA",
                                    "rekening": 57998977,
                                    "saldo": 50000
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Akan memberikan data user yang sudah terdaftar dan pesan berhasil mendaftar"
                    },
                    "422": {
                        "description": "Some variable is required"
                    },
                    "403": {
                        "description": "duplicate entry"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}